agent/requirements.txt
openai

python-dotenv
python-amazon-paapi
requests
